+++ src/index.ts
@@ -5,7 +5,7 @@ export { encodeNsite, decodeNsite, nsite } from './encoders/nsite';
 export { encodeNfilter, decodeNfilter, nfilter } from './encoders/nfilter';
 export { encodeNfilters, decodeNfilters, nfilters } from './encoders/nfilters';
 export { encodeNfeed, decodeNfeed, nfeed } from './encoders/nfeed';
-export { encodeNvite, decodeNvite, nvite } from './encoders/nvite';
+export { encodeNinvite, decodeNinvite, ninvite } from './encoders/ninvite';
 export { encodeNapp, decodeNapp, napp } from './encoders/napp';
 export { encodeNblob, decodeNblob, nblob } from './encoders/nblob';
 
@@ -14,7 +14,7 @@ import { nsite } from './encoders/nsite';
 import { nfilter } from './encoders/nfilter';
 import { nfilters } from './encoders/nfilters';
 import { nfeed } from './encoders/nfeed';
-import { nvite } from './encoders/nvite';
+import { ninvite } from './encoders/ninvite';
 import { napp } from './encoders/napp';
 import { nblob } from './encoders/nblob';
 
@@ -24,7 +24,7 @@ export default {
   nfilter,
   nfilters,
   nfeed,
-  nvite,
+  ninvite,
   napp,
   nblob
 };